+++ src/components/views/content-card/ContentCardLink.view.tsx
@@ -1,0 +1,58 @@
+import { type FC, type PropsWithChildren } from "react";
+import { COLORS, TRANSITIONS, APP_ADDRESS } from "_constants";
+import { motion } from "framer-motion";
+import c from "classnames";
+import Link from "next/link";
+
+type ContentCardLinkViewProps = PropsWithChildren<{
+  href: string;
+}>;
+
+const ContentCardLinkView: FC<ContentCardLinkViewProps> = ({
+  href,
+  children,
+}) => {
+  const isExternalLink = !href.startsWith(APP_ADDRESS) && !href.startsWith("/");
+  return (
+    <motion.div
+      className="mb-3 last:mb-0 rounded-lg"
+      layout
+      whileHover={{ backgroundColor: COLORS.theme }}
+      transition={TRANSITIONS.routeFast}
+    >
+      {isExternalLink ? (
+        <a
+          href={href}
+          target="_blank"
+          rel="noopener noreferrer"
+          className="cursor-pointer relative block"
+        >
+          <div
+            className={c(
+              COLORS.secondaryText,
+              "absolute top-0 right-0 px-5 py-4 h-10 v-5"
+            )}
+          >
+            <ExternalLinkIcon />
+          </div>
+          {children}
+        </a>
+      ) : (
+        <Link href={href} passHref>
+          <a>{children}</a>
+        </Link>
+      )}
+    </motion.div>
+  );
+};
+
+const ExternalLinkIcon = () => (
+  <div
+    dangerouslySetInnerHTML={{
+      __html: require("_assets/icons/external-link.svg?include"),
+    }}
+    className={c("w-4 h-4", COLORS.secondaryFill)}
+  />
+);
+
+export default ContentCardLinkView;
